Code cleanup ViewModels moved to ui features dirs
